requirements.txt
appdirs==1.4.0
Django==1.11.4
packaging==16.8
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
mysqlclient==1.3.7
django-mysql==1.0.1
mod-wsgi==4.5.14
pycrypto==2.6.1
djangorestframework==3.6.3
markdown==2.6.8
django-filter==1.0.4
pypng
pyqrcode
